check_parallel| Fixups imports after file-move
